clinker-core: extract source, output, scoped-var, and config-error modules
Continue the config/mod.rs decomposition by relocating the
medium-complexity clusters into focused submodules:

- source.rs     — SourceConfig and its impl, WatermarkConfig, file-listing
                  controls, the rest/file transport enums, and every
                  per-format input-options struct
- output.rs     — OutputConfig, the if-exists / split policies, and every
                  per-format output-options struct
- scoped_var.rs — scoped-variable declarations, the registry builder, and
                  the default-collection / validation helpers
- error.rs      — the ConfigError hierarchy, env-var interpolation, and the
                  parse_config / load_config entry points

Pure relocation: no symbol renames and no public-API change. Every
existing `clinker_core::config::*` path still resolves through the
re-exports. Two validation helpers shared across the new boundary
(`validate_config`, `validate_unique_scoped_declarations`) gain
`pub(crate)` visibility; nothing new becomes `pub`.

Closes #256
