+++ crates/clinker-core/src/config/error.rs
@@ -1,0 +1,164 @@
+//! Config loading, env-var interpolation, and the [`ConfigError`] hierarchy.
+
+use super::*;
+use regex::Regex;
+use std::sync::LazyLock;
+
+/// Errors during config loading and parsing.
+#[derive(Debug)]
+pub enum ConfigError {
+    Io(std::io::Error),
+    Yaml(crate::yaml::YamlError),
+    EnvVar { var_name: String, position: usize },
+    Validation(String),
+}
+
+impl std::fmt::Display for ConfigError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::Io(e) => write!(f, "config I/O error: {e}"),
+            Self::Yaml(e) => write!(f, "YAML parse error: {e}"),
+            Self::EnvVar { var_name, position } => {
+                write!(
+                    f,
+                    "undefined environment variable ${{{var_name}}} at position {position}"
+                )
+            }
+            Self::Validation(msg) => write!(f, "config validation error: {msg}"),
+        }
+    }
+}
+
+impl std::error::Error for ConfigError {}
+
+impl From<std::io::Error> for ConfigError {
+    fn from(e: std::io::Error) -> Self {
+        Self::Io(e)
+    }
+}
+
+impl From<crate::yaml::YamlError> for ConfigError {
+    fn from(e: crate::yaml::YamlError) -> Self {
+        Self::Yaml(e)
+    }
+}
+
+// Regex for ${VAR} and ${VAR:-default} interpolation
+static ENV_VAR_RE: LazyLock<Regex> =
+    LazyLock::new(|| Regex::new(r"\$\{([A-Za-z_][A-Za-z0-9_]*)(?::-(.*?))?\}").unwrap());
+
+/// Pre-deserialize environment variable interpolation.
+///
+/// Replaces `${VAR}` with `env::var("VAR")` and `${VAR:-default}` with
+/// the env value or the default. Bare text substitution — no YAML quoting
+/// applied (industry standard: dbt, Docker Compose, Helm, envsubst).
+///
+/// `extra_vars` are checked before `std::env::var` — highest priority among
+/// runtime vars. This allows channel variables to shadow system env vars.
+///
+/// `$${VAR}` produces literal `${VAR}` (escape, same as Docker Compose /
+/// Kubernetes convention). The `$$` is replaced with a NULL byte placeholder
+/// before the main regex, then restored after all substitution.
+///
+/// Missing var without default produces `ConfigError::EnvVar`.
+pub fn interpolate_env_vars(
+    yaml: &str,
+    extra_vars: &[(&str, &str)],
+) -> Result<String, ConfigError> {
+    // Step 1: Replace $$ with NULL byte placeholder before main regex.
+    // NULL bytes do not appear in valid YAML config files.
+    debug_assert!(
+        !yaml.contains('\0'),
+        "input YAML contains NULL bytes — $$ escape placeholder collision"
+    );
+    let escaped = yaml.replace("$$", "\0");
+
+    // Step 2: Run regex substitution with extra_vars priority
+    let mut result = String::with_capacity(escaped.len());
+    let mut last_end = 0;
+
+    for caps in ENV_VAR_RE.captures_iter(&escaped) {
+        let full_match = caps.get(0).unwrap();
+        let var_name = caps.get(1).unwrap().as_str();
+        let default_value = caps.get(2).map(|m| m.as_str());
+
+        // Validate env var name: must be UPPERCASE + underscores only
+        if !var_name
+            .chars()
+            .all(|c| c.is_ascii_uppercase() || c.is_ascii_digit() || c == '_')
+            || var_name.starts_with(|c: char| c.is_ascii_digit())
+            || var_name.is_empty()
+        {
+            return Err(ConfigError::Validation(format!(
+                "invalid environment variable name '{}' at position {} — must match [A-Z_][A-Z0-9_]*",
+                var_name,
+                full_match.start()
+            )));
+        }
+
+        result.push_str(&escaped[last_end..full_match.start()]);
+
+        // Check extra_vars first, then system env
+        if let Some(value) = extra_vars
+            .iter()
+            .find(|(k, _)| *k == var_name)
+            .map(|(_, v)| *v)
+        {
+            result.push_str(value);
+        } else {
+            match std::env::var(var_name) {
+                Ok(value) => result.push_str(&value),
+                Err(_) => match default_value {
+                    Some(default) => result.push_str(default),
+                    None => {
+                        return Err(ConfigError::EnvVar {
+                            var_name: var_name.to_string(),
+                            position: full_match.start(),
+                        });
+                    }
+                },
+            }
+        }
+
+        last_end = full_match.end();
+    }
+
+    result.push_str(&escaped[last_end..]);
+
+    // Step 3: Restore NULL byte placeholder → $
+    Ok(result.replace('\0', "$"))
+}
+
+/// Parse a pipeline config from a YAML string (after interpolation).
+///
+/// All YAML parsing flows through `crate::yaml::from_str`, the single
+/// chokepoint that owns the DoS-defense [`Budget`].
+pub fn parse_config(yaml: &str) -> Result<PipelineConfig, ConfigError> {
+    let interpolated = interpolate_env_vars(yaml, &[])?;
+    let config: PipelineConfig = crate::yaml::from_str(&interpolated)?;
+    validate_config(&config)?;
+    Ok(config)
+}
+
+/// Load and parse a pipeline config from a YAML file path with extra variables.
+/// `extra_vars` are checked before system env vars during interpolation.
+///
+/// Stamps `config.source_hash` with the BLAKE3 of the post-env-var
+/// interpolated YAML so the executor can resolve `{pipeline_hash}` tokens
+/// and write provenance sidecars without retaining the source bytes.
+pub fn load_config_with_vars(
+    path: &std::path::Path,
+    extra_vars: &[(&str, &str)],
+) -> Result<PipelineConfig, ConfigError> {
+    let yaml = std::fs::read_to_string(path)?;
+    let interpolated = interpolate_env_vars(&yaml, extra_vars)?;
+    let mut config: PipelineConfig = crate::yaml::from_str(&interpolated)?;
+    config.source_hash = *blake3::hash(interpolated.as_bytes()).as_bytes();
+    validate_config(&config)?;
+    Ok(config)
+}
+
+/// Load and parse a pipeline config from a YAML file path.
+pub fn load_config(path: &std::path::Path) -> Result<PipelineConfig, ConfigError> {
+    load_config_with_vars(path, &[])
+}
