+++ crates/clinker-core/src/config/mod.rs
@@ -1,10 +1,14 @@
 pub mod compile_context;
 pub mod composition;
+pub mod error;
 pub mod format;
 pub mod node_header;
+pub mod output;
 pub mod pipeline_node;
 pub mod route;
+pub mod scoped_var;
 pub mod sort;
+pub mod source;
 pub mod utils;
 
 pub use crate::plan::index::AnalyticWindowSpec;
@@ -15,24 +19,25 @@ pub use composition::{
     ResolvedValue, Resource, ResourceDecl, ResourceKind, ResourceName, ScopedVarsSchema, SourceMap,
     SpannedNodeRef, WORKSPACE_COMPOSITION_BUDGET, scan_workspace_signatures, validate_signatures,
 };
+pub use error::*;
 pub use format::*;
 pub use node_header::{MergeHeader, NodeHeader, NodeInput, SourceHeader};
+pub use output::*;
 pub use pipeline_node::{
     AggregateBody, MergeBody, OutputBody, Phase, PipelineNode, RouteBody, SourceBody,
     TransformBody, VarScope,
 };
 pub use route::*;
+pub use scoped_var::*;
 pub use sort::*;
+pub use source::*;
 pub use utils::*;
 
 use crate::yaml::Spanned;
-use clinker_record::schema_def::{FieldDef, LineSeparator};
 use indexmap::IndexMap;
-use regex::Regex;
 use serde::de::{self, Visitor};
 use serde::{Deserialize, Deserializer, Serialize};
 use std::collections::HashMap;
-use std::sync::LazyLock;
 
 /// Top-level pipeline configuration, deserialized from YAML.
 ///
@@ -219,804 +224,6 @@ pub struct ConcurrencyConfig {
     pub threads: Option<usize>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub chunk_size: Option<usize>,
-}
-
-/// One static-config variable's declared type and optional default value.
-///
-/// `default` must match `var_type` — enforced by [`validate_static_vars`].
-/// Static-config vars are read via `$vars.<key>`; producer-written scoped
-/// state (`$pipeline.*` / `$source.*` / `$record.*`) is declared on
-/// Transforms via their `declares:` block instead.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct ScopedVarDecl {
-    #[serde(rename = "type")]
-    pub var_type: ScopedVarType,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub default: Option<serde_json::Value>,
-}
-
-/// Scoped-variable primitive type set.
-///
-/// Mirrors the CXL primitive types the typecheck pass uses to validate
-/// `$pipeline.<key>` / `$source.<key>` / `$record.<key>` reads. `Date`
-/// and `DateTime` accept ISO-8601 string defaults (parsed at the eval
-/// boundary, not at YAML load).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum ScopedVarType {
-    String,
-    Int,
-    Float,
-    Bool,
-    Date,
-    DateTime,
-}
-
-impl From<ScopedVarType> for cxl::resolve::ScopedVarType {
-    fn from(t: ScopedVarType) -> Self {
-        match t {
-            ScopedVarType::String => cxl::resolve::ScopedVarType::String,
-            ScopedVarType::Int => cxl::resolve::ScopedVarType::Int,
-            ScopedVarType::Float => cxl::resolve::ScopedVarType::Float,
-            ScopedVarType::Bool => cxl::resolve::ScopedVarType::Bool,
-            ScopedVarType::Date => cxl::resolve::ScopedVarType::Date,
-            ScopedVarType::DateTime => cxl::resolve::ScopedVarType::DateTime,
-        }
-    }
-}
-
-/// Build the resolver-side scoped-vars registry from:
-/// (a) the flat top-level `vars:` block — frozen `$vars.<key>` config,
-///     channel-overridable;
-/// (b) every Transform's `config.declares:` entries — producer-declared
-///     scoped state addressable as `$pipeline.<key>` / `$source.<key>` /
-///     `$record.<key>`. The writer-site declaration is the sole authority.
-pub fn build_scoped_vars_registry(
-    vars: Option<&IndexMap<String, ScopedVarDecl>>,
-    nodes: &[crate::yaml::Spanned<crate::config::pipeline_node::PipelineNode>],
-) -> cxl::resolve::ScopedVarsRegistry {
-    use crate::config::pipeline_node::{PipelineNode, VarScope};
-    let mut reg = cxl::resolve::ScopedVarsRegistry {
-        pipeline: indexmap::IndexMap::new(),
-        source: indexmap::IndexMap::new(),
-        record: indexmap::IndexMap::new(),
-        // Top-level pipelines never have hidden vars — those are
-        // populated only when entering a composition body via
-        // `build_body_scoped_vars` to support the E173 diagnostic.
-        hidden_pipeline: indexmap::IndexMap::new(),
-        hidden_source: indexmap::IndexMap::new(),
-        hidden_record: indexmap::IndexMap::new(),
-        static_vars: vars
-            .map(|sv| {
-                sv.iter()
-                    .map(|(k, d)| (k.clone(), d.var_type.into()))
-                    .collect()
-            })
-            .unwrap_or_default(),
-    };
-    for spanned in nodes {
-        if let PipelineNode::Transform { config, .. } = &spanned.value {
-            for entry in &config.declares {
-                let target = match entry.scope {
-                    VarScope::Pipeline => &mut reg.pipeline,
-                    VarScope::Source => &mut reg.source,
-                    VarScope::Record => &mut reg.record,
-                };
-                target.insert(entry.name.clone(), entry.var_type.into());
-            }
-        }
-    }
-    reg
-}
-
-/// Input source configuration.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SourceConfig {
-    pub name: String,
-    /// Literal file path (the simple case). Mutually exclusive with
-    /// `glob`/`regex`/`paths`; exactly one of the four must be set.
-    /// Validated post-deserialize via `SourceConfig::validate_matching`.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub path: Option<String>,
-    /// Glob pattern matching one or more files (POSIX/gitignore semantics
-    /// via the `glob` crate). Mutually exclusive with `path`/`regex`/`paths`.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub glob: Option<String>,
-    /// Regex pattern matched against full paths under the workspace root.
-    /// Mutually exclusive with `path`/`glob`/`paths`.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub regex: Option<String>,
-    /// Explicit list of files. Mutually exclusive with `path`/`glob`/`regex`.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub paths: Option<Vec<String>>,
-
-    /// Glob patterns to remove from the discovered file set (gitignore
-    /// semantics). Applied after primary discovery.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub exclude: Option<Vec<String>>,
-    /// Skip files modified before this point. Accepts a duration relative
-    /// to "now" (`"5m"`, `"2h"`, `"3d"`) or an RFC3339 timestamp.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub modified_after: Option<TimeBound>,
-    /// Skip files modified after this point. Accepts the same forms as
-    /// `modified_after`.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub modified_before: Option<TimeBound>,
-    /// Skip files smaller than this size. Accepts `"1KB"`, `"10MB"`, etc.
-    /// (decimal, 1KB = 1000 bytes).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub min_size: Option<ByteSize>,
-    /// Skip files larger than this size.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub max_size: Option<ByteSize>,
-
-    /// File-listing controls (sort order, take N, recursion, no-match
-    /// policy). Nested to keep the file-level `sort_order` separate from
-    /// the existing record-level `sort_order` field below.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub files: Option<FileListingControls>,
-
-    /// Event-time watermark declaration. Names the column whose value
-    /// is observed at ingest and folded into a per-(source, file)
-    /// max-event-time stamp; rolled up at the read side by
-    /// `PerSourceWatermarks::min_across_sources`. Mirrors Flink SQL's
-    /// `WATERMARK FOR <col> AS <col> - INTERVAL` declaration site;
-    /// the `INTERVAL` delay column lands with the time-window operator
-    /// at https://github.com/rustpunk/clinker/issues/61.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub watermark: Option<WatermarkConfig>,
-
-    /// Envelope-section configuration for document-aware sources.
-    /// Declared sections are extracted by the reader's one-time
-    /// pre-scan and exposed to CXL via `$doc.<section>.<field>`.
-    /// Section names are user-chosen identifiers; the engine reserves
-    /// none. Absent (`None`) when the source has no envelope to extract.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub envelope: Option<clinker_format::EnvelopeConfig>,
-
-    /// Format-layer schema pointer (e.g. fixed-width field layouts).
-    /// Distinct from the CXL-type-level `SourceBody.schema` declared
-    /// at the parent `SourceBody` scope — this one points at on-disk
-    /// format metadata, the other declares column CXL types for
-    /// compile-time typecheck.
-    #[serde(rename = "format_schema", skip_serializing_if = "Option::is_none")]
-    pub schema: Option<SchemaSource>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub schema_overrides: Option<Vec<FieldDef>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub array_paths: Option<Vec<ArrayPathConfig>>,
-    /// Record-level sortedness inside the file (used by combine/aggregate
-    /// planning to enable streaming strategies). Distinct from
-    /// `files.sort_order` which orders the file set itself.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub sort_order: Option<Vec<SortFieldSpec>>,
-    /// Transport selecting WHERE the records come from, sitting above the
-    /// on-disk FORMAT (`format` / `type:` below). A plain optional
-    /// snake_case enum field — deliberately NOT a second
-    /// `#[serde(flatten)]`+tagged enum, because a flattened tagged enum
-    /// loses YAML source spans (the same edge case `InputFormat` already
-    /// hits), and a second one would compound that loss. Defaults to
-    /// [`SourceTransport::File`] so every existing file pipeline that
-    /// omits `transport:` parses byte-identically.
-    #[serde(default)]
-    pub transport: SourceTransport,
-    #[serde(flatten)]
-    pub format: InputFormat,
-    /// Kiln IDE metadata: stage notes + field annotations. Ignored by the engine.
-    #[serde(default, rename = "_notes", skip_serializing_if = "Option::is_none")]
-    pub notes: Option<serde_json::Value>,
-}
-
-/// Event-time watermark declaration on a `SourceConfig`.
-///
-/// `column` names a record field whose value is the source's event-
-/// time axis. At ingest, each record's value at this column is
-/// converted to an i64 nanosecond stamp, shifted earlier by `delay`,
-/// and folded into a per-(source, file) max via
-/// [`crate::executor::watermark::PerSourceWatermarks`]. The column
-/// type must coerce to [`clinker_record::Value::Timestamp`] or
-/// [`clinker_record::Value::Date`] (validated at plan time, see
-/// `crate::plan::bind_schema`).
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct WatermarkConfig {
-    /// Name of the event-time column on the source's declared schema.
-    pub column: String,
-    /// Duration the source's live mpsc receiver may stay quiet before
-    /// its partitions flip to [`crate::executor::watermark::WatermarkStatus::Idle`].
-    /// `None` (default) means "never go idle" — preserves prior behavior
-    /// for pipelines without a window-close consumer.
-    ///
-    /// YAML form: a duration string with one of the `s` / `m` / `h` /
-    /// `d` / `ms` unit suffixes, e.g. `"30s"`, `"500ms"`, `"5m"`. Same
-    /// parser as [`TimeBound`]'s relative form, extended with `ms`
-    /// for the sub-second cadences a streaming consumer needs.
-    #[serde(
-        default,
-        deserialize_with = "deserialize_optional_duration",
-        serialize_with = "serialize_optional_duration",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub idle_timeout: Option<std::time::Duration>,
-    /// Bounded out-of-order tolerance for this source. Each record's
-    /// event-time, before being folded into the watermark, is shifted
-    /// earlier by `delay`; equivalently, the source's effective
-    /// watermark trails its observed max event-time by this amount.
-    /// Mirrors Flink's `BoundedOutOfOrdernessWatermarks` and Beam's
-    /// `WithAllowedTimestampSkew` — a source-system property declared
-    /// once at the source, distinct from the operator-side
-    /// `allowed_lateness` knob.
-    ///
-    /// `None` (default) means no shift — the watermark advances with
-    /// the observed max. Same YAML duration form as `idle_timeout`.
-    #[serde(
-        default,
-        deserialize_with = "deserialize_optional_duration",
-        serialize_with = "serialize_optional_duration",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub delay: Option<std::time::Duration>,
-}
-
-/// File-listing controls — file-set ordering, take-N, recursion,
-/// no-match policy. Nested under `files:` to avoid colliding with the
-/// record-level `sort_order:` field at the SourceConfig top level.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-#[serde(deny_unknown_fields, default)]
-pub struct FileListingControls {
-    /// Sort key for the file set. Default `Name`.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub sort_by: Option<FileSortKey>,
-    /// Sort direction. Default `Asc`.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub sort_order: Option<SortDir>,
-    /// Take only the first N files after sort. Mutually exclusive with
-    /// `take_last`.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub take_first: Option<usize>,
-    /// Take only the last N files after sort. Mutually exclusive with
-    /// `take_first`.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub take_last: Option<usize>,
-    /// Walk directories recursively. Default: true if the glob contains
-    /// `**`, false otherwise.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub recursive: Option<bool>,
-    /// Behavior when no files match. Default `Error`.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub on_no_match: Option<NoMatchPolicy>,
-}
-
-/// Sort key for ordering the discovered file set.
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum FileSortKey {
-    /// Lexicographic sort on the full path.
-    #[default]
-    Name,
-    /// Sort by file creation time (`birthtime` on supporting filesystems;
-    /// falls back to mtime where unavailable).
-    Created,
-    /// Sort by last-modified time.
-    Modified,
-}
-
-/// Sort direction for the file set ordering.
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum SortDir {
-    #[default]
-    Asc,
-    Desc,
-}
-
-/// Behavior when discovery matches zero files.
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum NoMatchPolicy {
-    /// Fail the run with E216.
-    #[default]
-    Error,
-    /// Log a warning and continue with an empty record stream.
-    Warn,
-    /// Silently produce an empty record stream.
-    Skip,
-}
-
-/// Transport selecting where a source's records originate, layered ABOVE
-/// the on-disk [`InputFormat`]. `File` (the default) reads bytes from the
-/// filesystem and decodes them with the declared format; the `Rest`
-/// transport (a paginated HTTP cursor) yields rows without going through
-/// fs discovery or the file matchers and carries its own
-/// connection/pagination config inline.
-///
-/// Selected by a nested `transport:` value, NOT a `#[serde(flatten)]`+
-/// tagged enum on `SourceConfig`. A second flattened tagged enum on
-/// `SourceConfig` would lose YAML source spans (the same edge case
-/// [`InputFormat`] already hits); a nested `transport:` value keeps the
-/// span on every payload field. Tagged with an internal `kind:` so the
-/// network variant can carry a payload struct:
-///
-/// ```yaml
-/// transport:
-///   kind: rest
-///   url: https://api.example.com/v1/records
-///   pagination: { strategy: link_header }
-///   max_pages: 50
-/// ```
-///
-/// Defaults to [`SourceTransport::File`] so every existing file pipeline
-/// that omits `transport:` (or writes the bare `transport: file` scalar)
-/// parses byte-identically — the `untagged`-free default arm plus the
-/// `kind` tag accept both the bare-`file` scalar and the nested forms.
-#[derive(Debug, Clone, Default, PartialEq, Serialize)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-pub enum SourceTransport {
-    /// Read bytes from the filesystem; resolve the file set through the
-    /// discovery layer's `path`/`glob`/`regex`/`paths` matchers.
-    #[default]
-    File,
-    /// Paginated HTTP GET cursor. Issues GETs to cursor exhaustion or a
-    /// hard page/record cap, decoding each response body through the
-    /// declared on-disk [`InputFormat`] (`json`/`xml`).
-    Rest(RestSourceConfig),
-}
-
-impl<'de> Deserialize<'de> for SourceTransport {
-    fn deserialize<D: Deserializer<'de>>(d: D) -> Result<Self, D::Error> {
-        // Two accepted YAML shapes:
-        //   transport: file                  (bare scalar)
-        //   transport: { kind: rest, ... }   (internally-tagged map)
-        // serde's internally-tagged enums reject a bare scalar for a unit
-        // variant, so the scalar form is handled explicitly and the map
-        // form delegates to a private `kind`-tagged mirror.
-        #[derive(Deserialize)]
-        #[serde(tag = "kind", rename_all = "snake_case")]
-        enum Tagged {
-            File,
-            Rest(RestSourceConfig),
-        }
-
-        #[derive(Deserialize)]
-        #[serde(untagged)]
-        enum Either {
-            Scalar(String),
-            Tagged(Tagged),
-        }
-
-        match Either::deserialize(d)? {
-            Either::Scalar(s) if s == "file" => Ok(SourceTransport::File),
-            Either::Scalar(other) => Err(de::Error::custom(format!(
-                "unknown transport {other:?}; the bare scalar form only accepts `file` — \
-                 the rest transport uses a nested `transport: {{ kind: rest, … }}`"
-            ))),
-            Either::Tagged(Tagged::File) => Ok(SourceTransport::File),
-            Either::Tagged(Tagged::Rest(c)) => Ok(SourceTransport::Rest(c)),
-        }
-    }
-}
-
-impl SourceTransport {
-    /// Whether this transport reads from the filesystem (the only
-    /// transport that goes through fs discovery and the file matchers).
-    pub fn is_file(&self) -> bool {
-        matches!(self, SourceTransport::File)
-    }
-
-    /// Short lowercase transport name for diagnostics.
-    pub fn transport_name(&self) -> &'static str {
-        match self {
-            SourceTransport::File => "file",
-            SourceTransport::Rest(_) => "rest",
-        }
-    }
-}
-
-/// Pagination strategy for a [`RestSourceConfig`]. Determines how the
-/// reader advances from one page to the next and how it detects the last
-/// page. Every strategy is bounded by the source's hard page/record cap —
-/// finiteness is a HARD reader property, not a server promise.
-#[derive(Debug, Clone, PartialEq, Eq, Default, Serialize, Deserialize)]
-#[serde(tag = "strategy", rename_all = "snake_case")]
-pub enum RestPagination {
-    /// Single GET, no pagination. The response body is the whole result.
-    #[default]
-    None,
-    /// `?<offset_param>=N&<limit_param>=L`, advancing the offset by the
-    /// page size each request; stops when a page returns fewer than
-    /// `limit` records (or the cap trips).
-    Offset {
-        /// Query parameter carrying the row offset (default `offset`).
-        #[serde(default = "default_offset_param")]
-        offset_param: String,
-        /// Query parameter carrying the page size (default `limit`).
-        #[serde(default = "default_limit_param")]
-        limit_param: String,
-        /// Page size requested per GET.
-        limit: u32,
-    },
-    /// Cursor-token pagination: the reader reads a token from a JSON
-    /// pointer in each response and sends it back as a query parameter
-    /// on the next request; stops when the token field is absent/null.
-    CursorToken {
-        /// Query parameter carrying the continuation token.
-        cursor_param: String,
-        /// JSON pointer (RFC 6901) into the response body locating the
-        /// next-page token (e.g. `/meta/next_cursor`).
-        next_token_pointer: String,
-    },
-    /// RFC 5988 `Link` header pagination: the reader follows the URL in
-    /// the `rel="next"` link until no such link is present.
-    LinkHeader,
-}
-
-fn default_offset_param() -> String {
-    "offset".to_string()
-}
-
-fn default_limit_param() -> String {
-    "limit".to_string()
-}
-
-/// HTTP authentication for a [`RestSourceConfig`].
-#[derive(Debug, Clone, PartialEq, Eq, Default, Serialize, Deserialize)]
-#[serde(tag = "scheme", rename_all = "snake_case")]
-pub enum RestAuth {
-    /// No authentication header.
-    #[default]
-    None,
-    /// `Authorization: Bearer <token>`.
-    Bearer {
-        /// The bearer token, sent verbatim after `Bearer `.
-        token: String,
-    },
-    /// An arbitrary static header, e.g. `X-API-Key: <value>`.
-    Header {
-        /// Header name.
-        name: String,
-        /// Header value.
-        value: String,
-    },
-}
-
-/// Connection + pagination config for the `rest` transport. The reader
-/// issues paginated GETs against [`url`](Self::url) until the pagination
-/// strategy reports no next page OR a hard cap trips, decoding each
-/// response body through the source's declared on-disk format.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct RestSourceConfig {
-    /// Base request URL. Pagination appends/overrides query parameters.
-    pub url: String,
-    /// Pagination strategy (default [`RestPagination::None`] — a single
-    /// GET).
-    #[serde(default)]
-    pub pagination: RestPagination,
-    /// Authentication (default [`RestAuth::None`]).
-    #[serde(default)]
-    pub auth: RestAuth,
-    /// HARD page cap. The reader stops after this many GETs even if the
-    /// server keeps offering a next page. Required — finiteness is
-    /// unenforceable downstream, so an explicit ceiling is mandatory.
-    pub max_pages: u32,
-    /// Optional HARD record cap, applied in addition to `max_pages`. The
-    /// reader stops emitting once this many records have been yielded.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub max_records: Option<u64>,
-    /// Bounded transient-failure retry count per request (5xx / timeout /
-    /// connect error). Default 3.
-    #[serde(default = "default_rest_retries")]
-    pub retries: u32,
-    /// Per-request timeout in seconds. Bounds in-flight request time so
-    /// SIGINT cancellation lands within the documented shutdown bound.
-    /// Default 30.
-    #[serde(default = "default_rest_timeout_secs")]
-    pub timeout_secs: u64,
-}
-
-fn default_rest_retries() -> u32 {
-    3
-}
-
-fn default_rest_timeout_secs() -> u64 {
-    30
-}
-
-/// CSV-specific input options.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-#[serde(deny_unknown_fields, default)]
-pub struct CsvInputOptions {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub delimiter: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub quote_char: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub has_header: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub encoding: Option<String>,
-}
-
-/// JSON-specific input options.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-#[serde(deny_unknown_fields, default)]
-pub struct JsonInputOptions {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub format: Option<JsonFormat>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub record_path: Option<String>,
-}
-
-/// XML-specific input options.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-#[serde(deny_unknown_fields, default)]
-pub struct XmlInputOptions {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub record_path: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub attribute_prefix: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub namespace_handling: Option<NamespaceHandling>,
-}
-
-/// JSON input format mode (auto-detect if not specified).
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum JsonFormat {
-    Array,
-    Ndjson,
-    Object,
-}
-
-/// XML namespace handling mode.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum NamespaceHandling {
-    #[default]
-    Strip,
-    Qualify,
-}
-
-/// Array path configuration for nested array explosion/joining.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ArrayPathConfig {
-    pub path: String,
-    #[serde(default)]
-    pub mode: ArrayMode,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub separator: Option<String>,
-}
-
-/// Array path processing mode.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ArrayMode {
-    #[default]
-    Explode,
-    Join,
-}
-
-impl SourceConfig {
-    /// Get CSV input options, if this is a CSV input.
-    pub fn csv_options(&self) -> Option<&CsvInputOptions> {
-        match &self.format {
-            InputFormat::Csv(opts) => opts.as_ref(),
-            _ => None,
-        }
-    }
-
-    /// Borrow the literal `path:` for the simple-case source. Returns `""`
-    /// for sources using `glob`/`regex`/`paths` matchers — the latter
-    /// resolve to a `Vec<PathBuf>` at discovery time and don't have a
-    /// single literal path to display.
-    ///
-    /// Use this only on call sites that genuinely want the literal path
-    /// (display, security checks of one path, schema-stem extraction for
-    /// the simple case). For runtime file enumeration use the discovery
-    /// module instead.
-    pub fn path_str(&self) -> &str {
-        self.path.as_deref().unwrap_or("")
-    }
-
-    /// Human-readable target description for diagnostics and Kiln display.
-    /// Reflects the active matcher (`path` / `glob` / `regex` / `paths`).
-    pub fn display_target(&self) -> String {
-        if let Some(p) = self.path.as_deref() {
-            p.to_string()
-        } else if let Some(g) = self.glob.as_deref() {
-            format!("glob: {g}")
-        } else if let Some(r) = self.regex.as_deref() {
-            format!("regex: {r}")
-        } else if let Some(ps) = self.paths.as_deref() {
-            format!("paths: [{}]", ps.join(", "))
-        } else {
-            String::new()
-        }
-    }
-}
-
-/// Output destination configuration.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct OutputConfig {
-    pub name: String,
-    pub path: String,
-    /// When `true` (default), every input field not explicitly emitted
-    /// passes through to the output unchanged. When `false`, only
-    /// explicitly emitted fields appear in the output. Also the path
-    /// that surfaces `OnUnmapped::AutoWiden`-discovered columns at the
-    /// sink by expanding the `$widened` sidecar payload back to
-    /// top-level fields.
-    #[serde(default = "default_include_unmapped")]
-    pub include_unmapped: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub include_header: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub mapping: Option<IndexMap<String, String>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub exclude: Option<Vec<String>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub sort_order: Option<Vec<SortFieldSpec>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub preserve_nulls: Option<bool>,
-    /// Controls whether engine-stamped correlation snapshot columns
-    /// (`$ck.<field>`) appear in the default writer output. The shadow
-    /// columns preserve correlation-group identity through Transforms
-    /// that may rewrite the user-declared field; they are an internal
-    /// engine namespace and are stripped from output unless this flag
-    /// is set. Defaults to `false`.
-    #[serde(default)]
-    pub include_correlation_keys: bool,
-    /// Explicit schema for output formats that require field definitions
-    /// (e.g., fixed-width output needs field names, widths, and positions).
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub schema: Option<SchemaSource>,
-    /// File splitting configuration. When present, output is split into
-    /// multiple files based on record count or byte size limits.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub split: Option<SplitConfig>,
-    /// Optional per-Output override for the correlation fan-out policy.
-    /// Wins against the per-Combine override and the per-pipeline default
-    /// because the sink has the most context for whether collateral
-    /// rollback is acceptable (audit-style sinks typically opt down to
-    /// `Primary` or `All`; integrity-style sinks keep the default `Any`).
-    /// Additive opt-in: `None` defers to upstream resolution, preserving
-    /// today's behavior unchanged for every existing pipeline.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub correlation_fanout_policy: Option<CorrelationFanoutPolicy>,
-    /// Collision policy when the resolved output path already exists.
-    /// Defaults to `Overwrite` to preserve today's `File::create` behavior.
-    #[serde(default, skip_serializing_if = "IfExistsPolicy::is_default")]
-    pub if_exists: IfExistsPolicy,
-    /// Zero-pad width for the `{n}` collision counter when
-    /// `if_exists = unique_suffix`. `0` (default) emits the bare integer.
-    #[serde(default, skip_serializing_if = "is_zero_u8")]
-    pub unique_suffix_width: u8,
-    /// Write a `<resolved_path>.meta.json` provenance sidecar after the
-    /// output stream is flushed. Opt-in.
-    #[serde(default, skip_serializing_if = "is_false_bool")]
-    pub write_meta: bool,
-    #[serde(flatten)]
-    pub format: OutputFormat,
-    /// Kiln IDE metadata: stage notes + field annotations. Ignored by the engine.
-    #[serde(default, rename = "_notes", skip_serializing_if = "Option::is_none")]
-    pub notes: Option<serde_json::Value>,
-}
-
-/// Collision policy when an Output node's resolved path already exists.
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum IfExistsPolicy {
-    /// Truncate and overwrite. Today's `File::create` behavior.
-    #[default]
-    Overwrite,
-    /// Refuse to clobber; emit a diagnostic. `--force` downgrades to `Overwrite`.
-    Error,
-    /// Walk integer suffixes via `OpenOptions::create_new` until one succeeds.
-    UniqueSuffix,
-}
-
-impl IfExistsPolicy {
-    pub fn is_default(&self) -> bool {
-        matches!(self, IfExistsPolicy::Overwrite)
-    }
-}
-
-/// Output file splitting configuration.
-///
-/// When `group_key` is set, files are split only at key-group boundaries
-/// (greedy: first boundary after threshold). Without `group_key`, mechanical
-/// split at exact limit (like `split -l`).
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct SplitConfig {
-    /// Soft record count limit per file.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_records: Option<u64>,
-    /// Soft byte size limit per file.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_bytes: Option<u64>,
-    /// Optional key field — never split mid-group.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub group_key: Option<String>,
-    /// File naming pattern. Default: `"{stem}_{seq:04}.{ext}"`.
-    #[serde(default = "default_split_naming")]
-    pub naming: String,
-    /// CSV: repeat header row in each split file. Default: true.
-    #[serde(default = "default_true")]
-    pub repeat_header: bool,
-    /// Behavior when a single key group exceeds file limits.
-    #[serde(default)]
-    pub oversize_group: SplitOversizeGroupPolicy,
-}
-
-fn default_split_naming() -> String {
-    "{stem}_{seq:04}.{ext}".to_string()
-}
-
-/// Policy when a single key group exceeds split file limits.
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum SplitOversizeGroupPolicy {
-    /// Log a warning and allow the oversized file.
-    #[default]
-    Warn,
-    /// Error — pipeline stops.
-    Error,
-    /// Silently allow.
-    Allow,
-}
-
-/// CSV-specific output options.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-#[serde(deny_unknown_fields, default)]
-pub struct CsvOutputOptions {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub delimiter: Option<String>,
-}
-
-/// JSON-specific output options.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-#[serde(deny_unknown_fields, default)]
-pub struct JsonOutputOptions {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub format: Option<JsonOutputFormat>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub pretty: Option<bool>,
-}
-
-/// JSON output format mode.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum JsonOutputFormat {
-    #[default]
-    Array,
-    Ndjson,
-}
-
-/// XML-specific output options.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-#[serde(deny_unknown_fields, default)]
-pub struct XmlOutputOptions {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub root_element: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub record_element: Option<String>,
-}
-
-/// Fixed-width input options.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-#[serde(deny_unknown_fields, default)]
-pub struct FixedWidthInputOptions {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub line_separator: Option<LineSeparator>,
-}
-
-/// Fixed-width output options.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-#[serde(deny_unknown_fields, default)]
-pub struct FixedWidthOutputOptions {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub line_separator: Option<LineSeparator>,
 }
 
 /// Input wiring for a transform — specifies which upstream transform(s) feed records.
@@ -3534,142 +2741,6 @@ pub struct DlqPerSourceConfig {
 /// override is supplied.
 pub const DEFAULT_DLQ_MIN_RECORDS: u64 = 100;
 
-/// Errors during config loading and parsing.
-#[derive(Debug)]
-pub enum ConfigError {
-    Io(std::io::Error),
-    Yaml(crate::yaml::YamlError),
-    EnvVar { var_name: String, position: usize },
-    Validation(String),
-}
-
-impl std::fmt::Display for ConfigError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Io(e) => write!(f, "config I/O error: {e}"),
-            Self::Yaml(e) => write!(f, "YAML parse error: {e}"),
-            Self::EnvVar { var_name, position } => {
-                write!(
-                    f,
-                    "undefined environment variable ${{{var_name}}} at position {position}"
-                )
-            }
-            Self::Validation(msg) => write!(f, "config validation error: {msg}"),
-        }
-    }
-}
-
-impl std::error::Error for ConfigError {}
-
-impl From<std::io::Error> for ConfigError {
-    fn from(e: std::io::Error) -> Self {
-        Self::Io(e)
-    }
-}
-
-impl From<crate::yaml::YamlError> for ConfigError {
-    fn from(e: crate::yaml::YamlError) -> Self {
-        Self::Yaml(e)
-    }
-}
-
-// Regex for ${VAR} and ${VAR:-default} interpolation
-static ENV_VAR_RE: LazyLock<Regex> =
-    LazyLock::new(|| Regex::new(r"\$\{([A-Za-z_][A-Za-z0-9_]*)(?::-(.*?))?\}").unwrap());
-
-/// Pre-deserialize environment variable interpolation.
-///
-/// Replaces `${VAR}` with `env::var("VAR")` and `${VAR:-default}` with
-/// the env value or the default. Bare text substitution — no YAML quoting
-/// applied (industry standard: dbt, Docker Compose, Helm, envsubst).
-///
-/// `extra_vars` are checked before `std::env::var` — highest priority among
-/// runtime vars. This allows channel variables to shadow system env vars.
-///
-/// `$${VAR}` produces literal `${VAR}` (escape, same as Docker Compose /
-/// Kubernetes convention). The `$$` is replaced with a NULL byte placeholder
-/// before the main regex, then restored after all substitution.
-///
-/// Missing var without default produces `ConfigError::EnvVar`.
-pub fn interpolate_env_vars(
-    yaml: &str,
-    extra_vars: &[(&str, &str)],
-) -> Result<String, ConfigError> {
-    // Step 1: Replace $$ with NULL byte placeholder before main regex.
-    // NULL bytes do not appear in valid YAML config files.
-    debug_assert!(
-        !yaml.contains('\0'),
-        "input YAML contains NULL bytes — $$ escape placeholder collision"
-    );
-    let escaped = yaml.replace("$$", "\0");
-
-    // Step 2: Run regex substitution with extra_vars priority
-    let mut result = String::with_capacity(escaped.len());
-    let mut last_end = 0;
-
-    for caps in ENV_VAR_RE.captures_iter(&escaped) {
-        let full_match = caps.get(0).unwrap();
-        let var_name = caps.get(1).unwrap().as_str();
-        let default_value = caps.get(2).map(|m| m.as_str());
-
-        // Validate env var name: must be UPPERCASE + underscores only
-        if !var_name
-            .chars()
-            .all(|c| c.is_ascii_uppercase() || c.is_ascii_digit() || c == '_')
-            || var_name.starts_with(|c: char| c.is_ascii_digit())
-            || var_name.is_empty()
-        {
-            return Err(ConfigError::Validation(format!(
-                "invalid environment variable name '{}' at position {} — must match [A-Z_][A-Z0-9_]*",
-                var_name,
-                full_match.start()
-            )));
-        }
-
-        result.push_str(&escaped[last_end..full_match.start()]);
-
-        // Check extra_vars first, then system env
-        if let Some(value) = extra_vars
-            .iter()
-            .find(|(k, _)| *k == var_name)
-            .map(|(_, v)| *v)
-        {
-            result.push_str(value);
-        } else {
-            match std::env::var(var_name) {
-                Ok(value) => result.push_str(&value),
-                Err(_) => match default_value {
-                    Some(default) => result.push_str(default),
-                    None => {
-                        return Err(ConfigError::EnvVar {
-                            var_name: var_name.to_string(),
-                            position: full_match.start(),
-                        });
-                    }
-                },
-            }
-        }
-
-        last_end = full_match.end();
-    }
-
-    result.push_str(&escaped[last_end..]);
-
-    // Step 3: Restore NULL byte placeholder → $
-    Ok(result.replace('\0', "$"))
-}
-
-/// Parse a pipeline config from a YAML string (after interpolation).
-///
-/// All YAML parsing flows through `crate::yaml::from_str`, the single
-/// chokepoint that owns the DoS-defense [`Budget`].
-pub fn parse_config(yaml: &str) -> Result<PipelineConfig, ConfigError> {
-    let interpolated = interpolate_env_vars(yaml, &[])?;
-    let config: PipelineConfig = crate::yaml::from_str(&interpolated)?;
-    validate_config(&config)?;
-    Ok(config)
-}
-
 /// Reserved `$pipeline.*` member names that cannot be used as
 /// `declares:` entry names with `scope: pipeline`. Mirrors
 /// `crates/cxl/src/resolve/pass.rs::PIPELINE_MEMBERS`.
@@ -3909,213 +2980,6 @@ fn validate_config(config: &PipelineConfig) -> Result<(), ConfigError> {
     }
 
     Ok(())
-}
-
-/// Typecheck a serde-json default value against a declared scoped-var
-/// type. Used at config-load time for pipeline `vars:` and Transform
-/// `declares:` defaults, and at channel-bind time for channel var
-/// overrides — same invariant in every caller.
-pub fn check_scoped_var_default(
-    where_label: &str,
-    name: &str,
-    var_type: ScopedVarType,
-    default: &serde_json::Value,
-) -> Result<(), ConfigError> {
-    let mismatch = || {
-        ConfigError::Validation(format!(
-            "{where_label}.{name}: default value does not match declared type {var_type:?}",
-        ))
-    };
-    match (var_type, default) {
-        (_, serde_json::Value::Null) => Err(ConfigError::Validation(format!(
-            "{where_label}.{name}: default value cannot be null",
-        ))),
-        (
-            ScopedVarType::String | ScopedVarType::Date | ScopedVarType::DateTime,
-            serde_json::Value::String(_),
-        ) => Ok(()),
-        (ScopedVarType::Int, serde_json::Value::Number(n)) if n.is_i64() => Ok(()),
-        (ScopedVarType::Float, serde_json::Value::Number(_)) => Ok(()),
-        (ScopedVarType::Bool, serde_json::Value::Bool(_)) => Ok(()),
-        _ => Err(mismatch()),
-    }
-}
-
-/// Convert the flat `vars:` block into the runtime `$vars.*` value map.
-/// Entries without a `default` resolve to `Value::Null` at read time
-/// (the eval arm does the unwrap_or). Channels override these values
-/// once at pipeline start (Stage 7 wiring).
-pub fn convert_vars(
-    vars: &IndexMap<String, ScopedVarDecl>,
-) -> IndexMap<String, clinker_record::Value> {
-    vars.iter()
-        .filter_map(|(name, decl)| {
-            decl.default.as_ref().map(|default| {
-                (
-                    name.clone(),
-                    coerce_scoped_var_default(decl.var_type, default),
-                )
-            })
-        })
-        .collect()
-}
-
-/// Walk every Transform's `declares:` entries and collect declared
-/// defaults for variables whose `scope: pipeline`. Used at executor
-/// init to seed `StableEvalContext.pipeline_vars` before the runtime
-/// walk begins; reads of a declared `$pipeline.<key>` that fire
-/// before the writer (or whose writer never fires for a given record)
-/// resolve to the default rather than `Null`.
-pub fn collect_pipeline_var_defaults(
-    nodes: &[crate::yaml::Spanned<crate::config::pipeline_node::PipelineNode>],
-) -> IndexMap<String, clinker_record::Value> {
-    collect_scoped_var_defaults(nodes, pipeline_node::VarScope::Pipeline)
-}
-
-/// Walk every Transform's `declares:` entries and collect declared
-/// defaults for variables whose scope matches `wanted`. Pipeline,
-/// Source, and Record scopes share the same shape — pull them through
-/// the same helper rather than duplicating the loop body. Source and
-/// record defaults use this map as their pre-seed at executor init
-/// when channel-supplied overrides are absent.
-fn collect_scoped_var_defaults(
-    nodes: &[crate::yaml::Spanned<crate::config::pipeline_node::PipelineNode>],
-    wanted: pipeline_node::VarScope,
-) -> IndexMap<String, clinker_record::Value> {
-    use crate::config::pipeline_node::PipelineNode;
-    let mut out = IndexMap::new();
-    for spanned in nodes {
-        let PipelineNode::Transform { config, .. } = &spanned.value else {
-            continue;
-        };
-        for entry in &config.declares {
-            if entry.scope != wanted {
-                continue;
-            }
-            if let Some(default) = &entry.default {
-                out.insert(
-                    entry.name.clone(),
-                    coerce_scoped_var_default(entry.var_type, default),
-                );
-            }
-        }
-    }
-    out
-}
-
-/// Walk every Transform's `declares:` entries and collect declared
-/// defaults for variables whose `scope: source`. Channels overlay
-/// per-source-name overrides on top of this baseline.
-pub fn collect_source_var_defaults(
-    nodes: &[crate::yaml::Spanned<crate::config::pipeline_node::PipelineNode>],
-) -> IndexMap<String, clinker_record::Value> {
-    collect_scoped_var_defaults(nodes, pipeline_node::VarScope::Source)
-}
-
-/// Walk every Transform's `declares:` entries and collect declared
-/// defaults for variables whose `scope: record`. Channels overlay
-/// channel-wide defaults on top; the executor pre-seeds these into
-/// each record's `record_vars` map at materialization.
-pub fn collect_record_var_defaults(
-    nodes: &[crate::yaml::Spanned<crate::config::pipeline_node::PipelineNode>],
-) -> IndexMap<String, clinker_record::Value> {
-    collect_scoped_var_defaults(nodes, pipeline_node::VarScope::Record)
-}
-
-/// Walk the config's declared scopes ({Pipeline, Source, Record}) and
-/// reject duplicate names within each flat shared registry. Source-scope
-/// uniqueness applies globally (the runtime keys source-scope state by
-/// `Arc<str>` source-file but the declared-name namespace is shared).
-///
-/// A duplicate name implies ambiguity for downstream channel-overlay
-/// resolution and for the runtime read namespace itself — flat shared
-/// namespaces fail-fast (Beam, Flink, Kafka Streams, Dagster, Cargo,
-/// Rust statics, post-fix dbt). Authors who want shared state declare
-/// once and reference everywhere; clinker has no nested-scope construct
-/// for `$pipeline.*`/`$source.*`/`$record.*` that would justify
-/// shadowing semantics.
-fn validate_unique_scoped_declarations(config: &PipelineConfig) -> Result<(), ConfigError> {
-    use crate::config::pipeline_node::{PipelineNode, VarScope};
-    for scope in [VarScope::Pipeline, VarScope::Source, VarScope::Record] {
-        let mut first_seen: HashMap<&str, &str> = HashMap::new();
-        for spanned in &config.nodes {
-            let PipelineNode::Transform {
-                header,
-                config: body,
-            } = &spanned.value
-            else {
-                continue;
-            };
-            for entry in &body.declares {
-                if entry.scope != scope {
-                    continue;
-                }
-                if let Some(prior) = first_seen.insert(entry.name.as_str(), header.name.as_str())
-                    && prior != header.name.as_str()
-                {
-                    let scope_label = match scope {
-                        VarScope::Pipeline => "pipeline",
-                        VarScope::Source => "source",
-                        VarScope::Record => "record",
-                    };
-                    return Err(ConfigError::Validation(format!(
-                        "duplicate ${scope_label} declaration: '{name}' is declared in transforms '{prior}' and '{current}' — the {scope_label}-scope namespace is flat and shared, declare once and reference",
-                        name = entry.name,
-                        current = header.name,
-                    )));
-                }
-            }
-        }
-    }
-    Ok(())
-}
-
-/// Coerce a serde-json default value into a typed `clinker_record::Value`
-/// according to the declared scoped-var type. Caller MUST have run
-/// [`check_scoped_var_default`] first; this function panics on any
-/// (type, value) combo that validator would have rejected.
-pub fn coerce_scoped_var_default(
-    var_type: ScopedVarType,
-    default: &serde_json::Value,
-) -> clinker_record::Value {
-    use clinker_record::Value;
-    match (var_type, default) {
-        (ScopedVarType::Bool, serde_json::Value::Bool(b)) => Value::Bool(*b),
-        (ScopedVarType::Int, serde_json::Value::Number(n)) => {
-            Value::Integer(n.as_i64().unwrap_or(0))
-        }
-        (ScopedVarType::Float, serde_json::Value::Number(n)) => {
-            Value::Float(n.as_f64().unwrap_or(0.0))
-        }
-        (
-            ScopedVarType::String | ScopedVarType::Date | ScopedVarType::DateTime,
-            serde_json::Value::String(s),
-        ) => Value::String(s.clone().into_boxed_str()),
-        _ => unreachable!("check_scoped_var_default should reject mismatched defaults"),
-    }
-}
-
-/// Load and parse a pipeline config from a YAML file path with extra variables.
-/// `extra_vars` are checked before system env vars during interpolation.
-///
-/// Stamps `config.source_hash` with the BLAKE3 of the post-env-var
-/// interpolated YAML so the executor can resolve `{pipeline_hash}` tokens
-/// and write provenance sidecars without retaining the source bytes.
-pub fn load_config_with_vars(
-    path: &std::path::Path,
-    extra_vars: &[(&str, &str)],
-) -> Result<PipelineConfig, ConfigError> {
-    let yaml = std::fs::read_to_string(path)?;
-    let interpolated = interpolate_env_vars(&yaml, extra_vars)?;
-    let mut config: PipelineConfig = crate::yaml::from_str(&interpolated)?;
-    config.source_hash = *blake3::hash(interpolated.as_bytes()).as_bytes();
-    validate_config(&config)?;
-    Ok(config)
-}
-
-/// Load and parse a pipeline config from a YAML file path.
-pub fn load_config(path: &std::path::Path) -> Result<PipelineConfig, ConfigError> {
-    load_config_with_vars(path, &[])
 }
 
 /// Auto-extend every strict `PlanNode::Aggregation.group_by` with the
